requirements.txt
anthropic
PyMuPDF
Pillow
openpyxl
python-dotenv
fastapi
uvicorn
python-multipart
jinja2
